Refactor EditMetricPopup into screen-based dialog
